use hash map for better performance

+++ src/ds/include/ds/seek_two.h
@@ -6,7 +6,7 @@
 
 typedef struct int_tuple int_tuple;
 
-int_tuple* seek_two_count(size_t len, int16_t a[len], int16_t target);
+int_tuple* seek_two_count(size_t len, int32_t a[len], int32_t target);
 
 size_t seek_two_tuple_one(int_tuple* t);
 size_t seek_two_tuple_two(int_tuple* t);
